+++ spec/spec_helper.rb
@@ -27,7 +27,7 @@ ActiveRecord::Base.establish_connection
 # Create tables
 ActiveRecord::Schema.define do
   create_table :beta_feature_settings, force: true do |t|
-    t.integer :betable_id, null: false
+    t.bigint :betable_id, null: false
     t.string :betable_type, null: false
     t.string :betas, array: true, default: [], null: false
     t.timestamps
